+++ src/application/services/shortener.service.ts
@@ -1,6 +1,6 @@
 import { customAlphabet } from 'nanoid'
 
-import { NotFoundError } from '@shared/errors/not-found.error.js'
+import { NotFoundError } from '@presentation/errors/not-found.error.js'
 import { CodeExistsError } from '@infrastructure/errors/code-exists.error.js'
 import { IUrlRepository } from '@domain/repositories/url-repository.interface.js'
 import { ValidUrl } from '@domain/value-objects/valid-url.js'
