+++ src/presentation/docs/patchPaths.ts
