+++ src/presentation/errors/error-types.ts
