+++ src/presentation/routes/shortener.routes.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express'
 
-import { ShortenerController } from '../controllers/shortener.controller.js'
+import { ShortenerController } from '@presentation/controllers/shortener.controller.js'
 
 /**
  * Creates a router for URL shortening operations.
